Node addresses are now 64 bits.

+++ vm/defs.hpp
@@ -12,7 +12,7 @@ typedef uint32_t uint_val;
 typedef int32_t int_val;
 typedef double float_val;
 typedef uint64_t ptr_val;
-typedef uint32_t node_val;
+typedef uint64_t node_val;
 typedef bool bool_val;
 typedef unsigned char predicate_id;
 typedef unsigned short process_id;
